make scalac options very strict
fix and cleanup code to eliminate the resulting warnings/errors.
