+++ core/src/main/scala/datomisca/schema.scala
@@ -16,14 +16,10 @@
 
 package datomisca
 
-import scala.language.reflectiveCalls
-
 import java.{lang => jl}
 import java.math.{BigInteger => JBigInt, BigDecimal => JBigDecimal}
 import java.util.{Date, UUID}
 import java.net.URI
-
-import clojure.{lang => clj}
 
 
 object DatomicRef
